qemu better support
This patches copies all needed ppc files to qemu and modifies them.

The generated OpenBIOS image must be loaded at adresse 0xfff00000 by  Qemu.

It is able to load Yaboot from disk and Yaboot is able to load Linux, but Linux hangs somewhere... of course it needs more work (but this cleans up my patch queue).

Thank you to Blue Swirl and Aurelien Jarno for their help.



git-svn-id: svn://coreboot.org/openbios/openbios-devel@301 f158a5a8-5612-0410-a976-696ce0be7e32

+++ include/ofmem.h
@@ -27,6 +27,6 @@ extern ulong 	ofmem_claim_virt( ulong mvirt, ulong size, ulong align );
 extern int 	ofmem_map( ulong phys, ulong virt, ulong size, int mode );
 
 extern void  	ofmem_release( ulong virt, ulong size );
-extern ulong 	ofmem_translate( ulong virt, ulong *ret_mode );
+extern ulong 	ofmem_translate( ulong virt, int *ret_mode );
 
 #endif   /* _H_OFMEM */
